+++ src/app/features/pages/backoffice/dashboard/dashboard.component.scss
@@ -1,4 +1,5 @@
 .wrapper {
+  padding-top: 30px;
   width: 95%;
   justify-items: center;
   max-width: 1300px;
@@ -7,8 +8,13 @@
   grid-template-columns: 1fr 1fr 1fr 1fr;
   row-gap: 30px;
 }
-.card-grid {
+.card {
   text-align: center;
+  transition: all 0.2s ease-in-out;
+}
+.card:hover {
+  transform: scale(1.1);
+  cursor: pointer;
 }
 
 @media screen and (max-width: 1100px) {
