--- src/app/features/pages/dashboard/dashboard.component.html
@@ -1,7 +1,0 @@
-<div class="wrapper">
-    <div *ngFor="let category of categories" class="card-grid">
-        <p-card header="{{category}}" [style]="{'width':'250px', 'height':'250px'}">
-            <img src="" alt="">
-        </p-card>
-    </div>
-</div>
