--- src/app/features/pages/dashboard/dashboard.component.ts
@@ -1,23 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-
-@Component({
-  selector: "app-dashboard",
-  templateUrl: "./dashboard.component.html",
-  styleUrls: ["./dashboard.component.scss"],
-})
-export class DashboardComponent implements OnInit {
-  categories: string[] = [
-    "Novedades",
-    "Actividades",
-    "Categorias",
-    "Testimonios",
-    "Organizacion",
-    "Slides",
-    "Usuarios",
-    "Miembros",
-  ];
-
-  constructor() {}
-
-  ngOnInit(): void {}
-}
